% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_regex_pattern_set}
\alias{wafregional_get_regex_pattern_set}
\title{Returns the RegexPatternSet specified by RegexPatternSetId}
\usage{
wafregional_get_regex_pattern_set(RegexPatternSetId)
}
\arguments{
\item{RegexPatternSetId}{[required] The \code{RegexPatternSetId} of the RegexPatternSet that you want to get.
\code{RegexPatternSetId} is returned by CreateRegexPatternSet and by
ListRegexPatternSets.}
}
\description{
Returns the RegexPatternSet specified by \code{RegexPatternSetId}.
}
\section{Request syntax}{
\preformatted{wafregional$get_regex_pattern_set(
  RegexPatternSetId = "string"
)
}
}

\keyword{internal}
